model-backend/requirements.txt
Flask
tensorflow
numpy
pillow
h5py